Move test closer to its subject
