+++ contexts/logging-grpc-context/src/test/kotlin/io/spine/logging/context/grpc/GrpcContextDataProviderSpec.kt
@@ -27,10 +27,10 @@
 package io.spine.logging.context.grpc
 
 import io.spine.logging.flogger.context.ContextDataProvider
-import com.google.common.flogger.testing.AbstractContextDataProviderSpec
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.optional.shouldBePresent
 import io.kotest.matchers.types.shouldBeInstanceOf
+import io.spine.logging.flogger.context.AbstractContextDataProviderSpec
 import java.util.*
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
