+++ flogger/flogger-api/src/test/kotlin/io/spine/logging/flogger/context/AbstractContextDataProviderSpec.kt
@@ -24,16 +24,16 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.google.common.flogger.testing
+package io.spine.logging.flogger.context
 
+import com.google.common.flogger.testing.ConfigurableLogger
+import com.google.common.flogger.testing.FakeLoggerBackend
+import com.google.common.flogger.testing.shouldBeEmpty
+import com.google.common.flogger.testing.shouldContainInOrder
+import com.google.common.flogger.testing.shouldHaveSize
+import com.google.common.flogger.testing.shouldUniquelyContain
 import io.spine.logging.flogger.LogContext.Key
 import io.spine.logging.flogger.backend.Metadata
-import io.spine.logging.flogger.context.ContextDataProvider
-import io.spine.logging.flogger.context.LogLevelMap
-import io.spine.logging.flogger.context.ScopeType
-import io.spine.logging.flogger.context.ScopedLoggingContext
-import io.spine.logging.flogger.context.ScopedLoggingContexts
-import io.spine.logging.flogger.context.Tags
 import io.spine.logging.flogger.repeatedKey
 import io.spine.logging.flogger.singleKey
 import io.kotest.matchers.booleans.shouldBeFalse
@@ -58,9 +58,8 @@ private typealias LoggerName = String
  * @see <a href="https://github.com/google/flogger/blob/70c5aea863952ee61b3d33afb41f2841b6d63455/api/src/test/java/com/google/common/flogger/testing/AbstractScopedLoggingContextTest.java">
  *     Original Java code of Google Flogger</a>
  */
-@Suppress("FunctionName", "ClassName") // Tests aren't recognized in `main` sources.
 @DisplayName("`ContextDataProvider` should") // This name is to be overridden by inheritors.
-public abstract class AbstractContextDataProviderSpec {
+abstract class AbstractContextDataProviderSpec {
 
     private lateinit var contextData: ContextDataProvider
     private lateinit var context: ScopedLoggingContext
@@ -95,7 +94,7 @@ public abstract class AbstractContextDataProviderSpec {
      * Such is reported as a compiler warning.
      */
     @BeforeEach
-    public fun setImplementation() {
+    fun setImplementation() {
         contextData = implementationUnderTest
         context = contextData.getContextApiSingleton()
     }
@@ -121,11 +120,11 @@ public abstract class AbstractContextDataProviderSpec {
     }
 
     @Nested
-    public inner class
+    inner class
     `create a new context` {
 
         @Test
-        public fun `with tags`() {
+        fun `with tags`() {
             val tags = Tags.of("foo", "bar")
             contextTags.shouldBeEmpty()
             context.newContext()
@@ -139,7 +138,7 @@ public abstract class AbstractContextDataProviderSpec {
         }
 
         @Test
-        public fun `with metadata`() {
+        fun `with metadata`() {
             val (key, value) = FOO to "foo"
             contextMetadata.shouldBeEmpty()
             context.newContext()
@@ -154,7 +153,7 @@ public abstract class AbstractContextDataProviderSpec {
         }
 
         @Test
-        public fun `with a log level map`() {
+        fun `with a log level map`() {
             val defaultLevel = Level.FINE
             val mapping = "foo.bar" to defaultLevel
             val levelMap = LogLevelMap.create(mapOf(mapping), defaultLevel)
@@ -171,7 +170,7 @@ public abstract class AbstractContextDataProviderSpec {
         }
 
         @Test
-        public fun `with merged tags`() {
+        fun `with merged tags`() {
             val (name, value1, value2) = listOf("foo", "bar", "baz")
             val outerTags = Tags.of(name, value1)
             contextTags.shouldBeEmpty()
@@ -206,7 +205,7 @@ public abstract class AbstractContextDataProviderSpec {
          */
         @Test
         @Suppress("MagicNumber") // The assertion is readable without a constant.
-        public fun `with concatenated metadata`() {
+        fun `with concatenated metadata`() {
             val outerFoo = "outer-foo"
             val outerBar = "outer-bar"
 
@@ -247,7 +246,7 @@ public abstract class AbstractContextDataProviderSpec {
         }
 
         @Test
-        public fun `with merged level maps`() {
+        fun `with merged level maps`() {
             // Although, a logger name can be any `string`,
             // it is usually the name of a package or a class.
             val (other, fooBar, fooBarBaz) = listOf("other.package", "foo.bar", "foo.bar.Baz")
@@ -287,7 +286,7 @@ public abstract class AbstractContextDataProviderSpec {
         }
 
         @Test
-        public fun `with bound scope types`() {
+        fun `with bound scope types`() {
             contextData.getScope(SUB_TASK).shouldBeNull()
             contextData.getScope(BATCH_JOB).shouldBeNull()
 
@@ -331,7 +330,7 @@ public abstract class AbstractContextDataProviderSpec {
      * returned by [implementationUnderTest].
      */
     @Test
-    public fun `merge scope and log site tags`() {
+    fun `merge scope and log site tags`() {
         val backend = FakeLoggerBackend()
         val logger = ConfigurableLogger.create(backend)
         val logSiteTags = Tags.of("foo", "bar")
@@ -355,7 +354,7 @@ public abstract class AbstractContextDataProviderSpec {
     }
 
     @Test
-    public fun `not create a new scope instance if the same type is bound twice`() {
+    fun `not create a new scope instance if the same type is bound twice`() {
         contextData.getScope(SUB_TASK).shouldBeNull()
         context.newContext(SUB_TASK)
             .run {
